@@ -4,14 +4,12 @@ namespace App\Models;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Laravel\Sanctum\HasApiTokens;
 
 class User extends Authenticatable
 {
-    use HasApiTokens, HasFactory, Notifiable;
+    use HasFactory, Notifiable;
 
     /**
      * The attributes that are mass assignable.
@@ -22,7 +20,6 @@ class User extends Authenticatable
         'name',
         'email',
         'password',
-        'role',
     ];
 
     /**
@@ -36,51 +33,15 @@ class User extends Authenticatable
     ];
 
     /**
-     * The attributes that should be cast.
+     * Get the attributes that should be cast.
      *
-     * @var array<string, string>
+     * @return array<string, string>
      */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-        'password' => 'hashed',
-    ];
-
-    const ROLE_USER = 'user';
-    const ROLE_ADMIN = 'admin';
-
-    public static function roles(): array
+    protected function casts(): array
     {
         return [
-            self::ROLE_USER,
-            self::ROLE_ADMIN,
+            'email_verified_at' => 'datetime',
+            'password' => 'hashed',
         ];
-    }
-
-    /**
-     * Check if the user has a specific role.
-     *
-     * @param string|array $role The role(s) to check.
-     * @return bool Returns true if the user has the specified role(s), false otherwise.
-     */
-    public function hasRole(string|array $role): bool
-    {
-        return in_array($this->role, (array) $role);
-    }
-    /**
-     * Set the user's last activity date.
-     *
-     * @return void
-     */
-    public function setLastActivity(): void
-    {
-        Model::withoutTimestamps(function () {
-            $this->last_activity = now();
-            $this->save();
-        });
-    }
-
-    protected static function boot(): void
-    {
-        parent::boot();
     }
 }
